create palette sub folder in theme
